requirements.txt
boto==2.49.0
boto3==1.26.111
clip==0.2.0
datasets==2.13.1
faiss_cpu==1.7.3
gensim==4.3.1
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.2
pandas==2.0.3
Pillow==10.0.0
progress==1.6
pyarrow==12.0.1
Requests==2.31.0
scikit_learn==1.2.2
sentence_transformers==2.2.2
tabulate==0.9.0
torch==2.0.0
transformers==4.27.4
wget==3.2
